backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
et-xmlfile==1.1.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.3
MarkupSafe==2.1.5
openpyxl==3.1.3
pipenv==2023.12.1
platformdirs==4.2.1
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
soupsieve==2.5
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.26.1
Werkzeug==3.0.3
WTForms==3.1.2